# vocabularity.frontend/src/api/models/ActionResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ActionResponse = {
    error: string;
    status: number;
};