# vocabularity.frontend/src/api/models/AddWordsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddWordsResponse = {
    added: number;
    total: number;
};