# vocabularity.frontend/src/api/models/PageAction.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PageAction = {
    action_id: string;
    page_id: string;
};