# vocabularity.frontend/src/api/models/SuggestCandidatesAction.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WordDto } from './WordDto';
export type SuggestCandidatesAction = {
    action_id: string;
    words: Array<WordDto>;
};