# vocabularity.frontend/src/api/models/WordDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WordDto = {
    id: string;
    word: string;
    translation: string;
    transcription: string;
    examples: Array<string>;
    level: string;
    context_sentence: string;
};