# vocabularity.frontend/src/api/models/WordListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WordDto } from './WordDto';
export type WordListResponse = {
    words: Array<WordDto>;
    total: number;
    page: number;
    page_size: number;
};