# vocabularity.frontend/src/api/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/actions/":{"post":{"tags":["actions"],"summary":"Action","operationId":"action_actions__post","requestBody":{"content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/PageAction"},{"$ref":"#/components/schemas/SuggestCandidatesAction"},{"$ref":"#/components/schemas/ClearCandidatesAction"}],"title":"Body","discriminator":{"propertyName":"action_id","mapping":{"set_page":"#/components/schemas/PageAction","suggest_candidates":"#/components/schemas/SuggestCandidatesAction","clear_candidates":"#/components/schemas/ClearCandidatesAction"}}}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dictionary/words":{"post":{"tags":["dictionary"],"summary":"Add Words To Dictionary","operationId":"add_words_to_dictionary_dictionary_words_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WordDto"},"title":"Items"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddWordsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["dictionary"],"summary":"Delete Words From Dictionary","operationId":"delete_words_from_dictionary_dictionary_words_delete","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"title":"Ids"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteWordsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["dictionary"],"summary":"Get Words","operationId":"get_words_dictionary_words_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Page Size"}},{"name":"include_translate","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Include Translate"}},{"name":"include_transcription","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Include Transcription"}},{"name":"include_examples","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Include Examples"}}],"requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"title":"Ids"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WordListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ActionResponse":{"properties":{"error":{"type":"string","title":"Error"},"status":{"type":"integer","title":"Status"}},"type":"object","required":["error","status"],"title":"ActionResponse"},"AddWordsResponse":{"properties":{"added":{"type":"integer","title":"Added"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["added","total"],"title":"AddWordsResponse"},"ClearCandidatesAction":{"properties":{"action_id":{"type":"string","const":"clear_candidates","title":"Action Id"}},"type":"object","required":["action_id"],"title":"ClearCandidatesAction"},"DeleteWordsResponse":{"properties":{"deleted":{"type":"integer","title":"Deleted"}},"type":"object","required":["deleted"],"title":"DeleteWordsResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PageAction":{"properties":{"action_id":{"type":"string","const":"set_page","title":"Action Id"},"page_id":{"type":"string","title":"Page Id"}},"type":"object","required":["action_id","page_id"],"title":"PageAction"},"SuggestCandidatesAction":{"properties":{"action_id":{"type":"string","const":"suggest_candidates","title":"Action Id"},"words":{"items":{"$ref":"#/components/schemas/WordDto"},"type":"array","title":"Words"}},"type":"object","required":["action_id","words"],"title":"SuggestCandidatesAction"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WordDto":{"properties":{"id":{"type":"string","title":"Id"},"word":{"type":"string","title":"Word"},"translation":{"type":"string","title":"Translation"},"transcription":{"type":"string","title":"Transcription"},"examples":{"items":{"type":"string"},"type":"array","title":"Examples"},"level":{"type":"string","title":"Level"},"context_sentence":{"type":"string","title":"Context Sentence"}},"type":"object","required":["id","word","translation","transcription","examples","level","context_sentence"],"title":"WordDto"},"WordListResponse":{"properties":{"words":{"items":{"$ref":"#/components/schemas/WordDto"},"type":"array","title":"Words"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"}},"type":"object","required":["words","total","page","page_size"],"title":"WordListResponse"}}}}